fixed tests using pytest
