+++ .env
@@ -1,0 +1,2 @@
+
+PYTHONPATH=./src:./tests
