+++ src/monocle_apptrace/exporters/aws/s3_exporter.py
@@ -30,8 +30,8 @@ class S3SpanExporter(SpanExporterBase):
         self.export_interval = 1
         self.s3_client = boto3.client(
             's3',
-            aws_access_key_id=os.getenv('AWS_ACCESS_KEY_ID'),
-            aws_secret_access_key=os.getenv('AWS_SECRET_ACCESS_KEY'),
+            aws_access_key_id=os.getenv('AWS_ACCESS_KEY_ID_EXPORTER'),
+            aws_secret_access_key=os.getenv('AWS_SECRET_ACCESS_KEY_EXPORTER'),
             region_name=region_name,
         )
         self.bucket_name = bucket_name or os.getenv('MONOCLE_S3_BUCKET_NAME','default-bucket')
