+++ tests/__init__.py
@@ -1,15 +1,3 @@
-import os
-os.environ.setdefault("AZURE_OPENAI_API_DEPLOYMENT", "")
-os.environ.setdefault("AZURE_OPENAI_API_KEY", "")
-os.environ.setdefault("AZURE_OPENAI_API_VERSION", "")
-os.environ.setdefault("AZURE_OPENAI_ENDPOINT", "")
-os.environ.setdefault("OPENAI_API_KEY", "")
-os.environ.setdefault("MONOCLE_BLOB_CONNECTION_STRING", "")
-os.environ.setdefault("MONOCLE_BLOB_CONTAINER_NAME", "")
-os.environ.setdefault("OPENSEARCH_ENDPOINT_URL", "")
-os.environ.setdefault("AWS_ACCESS_KEY_ID", "")
-os.environ.setdefault("AWS_SECRET_ACCESS_KEY", "")
-os.environ.setdefault("MISTRAL_API_KEY", "")
