+++ tests/integration/test_bedrock_opensearch.py
@@ -1,26 +1,67 @@
-from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
-from monocle_apptrace.instrumentation.common.instrumentor import setup_monocle_telemetry
-from monocle_apptrace.instrumentation.common.instrumentor import set_context_properties
-from monocle.tests.common.custom_exporter import CustomConsoleSpanExporter
-custom_exporter = CustomConsoleSpanExporter()
-setup_monocle_telemetry(
-    workflow_name="bedrock_workflow",
-    span_processors=[BatchSpanProcessor(custom_exporter)],
-    wrapper_methods=[]
-)
 #Continue with code
-import boto3
-from langchain_community.vectorstores import OpenSearchVectorSearch
-from langchain_aws import BedrockEmbeddings
-from opensearchpy import RequestsHttpConnection
-from requests_aws4auth import AWS4Auth
-from botocore.exceptions import ClientError
 import os
-#opensearch endpoint url
 
-def produce_response(query):
+import boto3
+import pytest
+from botocore.exceptions import ClientError
+from common.custom_exporter import CustomConsoleSpanExporter
+from langchain_aws import BedrockEmbeddings
+from langchain_community.vectorstores import OpenSearchVectorSearch
+from opensearchpy import RequestsHttpConnection
+from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
+from requests_aws4auth import AWS4Auth
+
+from monocle_apptrace.instrumentation.common.instrumentor import (
+    set_context_properties,
+    setup_monocle_telemetry,
+)
+
+custom_exporter = CustomConsoleSpanExporter()
+
+@pytest.fixture(scope="module")
+def setup():
+    setup_monocle_telemetry(
+        workflow_name="bedrock_workflow",
+        span_processors=[BatchSpanProcessor(custom_exporter)],
+        wrapper_methods=[]
+    )
+
+@pytest.mark.integration()
+def test_bedrock_opensearch(setup):
+    query = "how?"
     similar_documents = search_similar_documents_opensearch(query)
-    return produce_llm_response(query, similar_documents)
+    produce_llm_response(query, similar_documents)
+
+    spans = custom_exporter.get_captured_spans()
+
+    for span in spans:
+        span_attributes = span.attributes
+        if 'span.type' in span_attributes and span_attributes["span.type"] == "retrieval":
+            # Assertions for all retrieval attributes
+            assert span_attributes["entity.1.name"] == "bedrock_workflow"
+            assert span_attributes["entity.1.type"] == "workflow.langchain"
+            assert span_attributes["entity.2.name"] == "OpenSearchVectorSearch"
+            assert span_attributes["entity.2.type"] == "vectorstore.OpenSearchVectorSearch"
+            assert "entity.2.deployment" in span_attributes
+            assert span_attributes["entity.3.name"] == "amazon.titan-embed-text-v1"
+            assert span_attributes["entity.3.type"] == "model.embedding.amazon.titan-embed-text-v1"
+
+        if 'span.type' in span_attributes and span_attributes["span.type"] == "inference":
+            # Assertions for all inference attributes
+            assert span_attributes["entity.1.name"] == "bedrock_workflow"
+            assert span_attributes["entity.1.type"] == "workflow.generic"
+            assert span_attributes["entity.2.type"] == "inference.aws_sagemaker"
+            assert "entity.2.inference_endpoint" in span_attributes
+            assert span_attributes["entity.3.name"] == "ai21.j2-mid-v1"
+            assert span_attributes["entity.3.type"] == "model.llm.ai21.j2-mid-v1"
+
+
+            # Assertions for metadata
+            span_input, span_output, span_metadata = span.events
+            assert "completion_tokens" in span_metadata.attributes
+            assert "prompt_tokens" in span_metadata.attributes
+            assert "total_tokens" in span_metadata.attributes
+
 
 def produce_llm_response(query,similar_documents):
     # Bedrock Client Setup
@@ -65,7 +106,7 @@ def build_context(similar_documents):
                "and say that you don't know the answer if this is not something you can answer on your own."
 
 def search_similar_documents_opensearch(query):
-    opensearch_url = os.environ['OPENSEARCH_ENDPOINT_URL']
+    opensearch_url = os.environ['OPENSEARCH_ENDPOINT_URL_BOTO']
     index_name = "embeddings-bedrock"  # Your index name
 
     bedrock_embeddings = BedrockEmbeddings(region_name="us-east-1",model_id="amazon.titan-embed-text-v1")
@@ -95,41 +136,6 @@ def search_similar_documents_opensearch(query):
     docs = retriever.get_relevant_documents(query)
     print(f"Retrieved docs: {docs}")
     return [doc.page_content for doc in docs]
-
-
-
-produce_response("how?")
-
-
-spans = custom_exporter.get_captured_spans()
-
-for span in spans:
-    span_attributes = span.attributes
-    if 'span.type' in span_attributes and span_attributes["span.type"] == "retrieval":
-        # Assertions for all retrieval attributes
-        assert span_attributes["entity.1.name"] == "bedrock_workflow"
-        assert span_attributes["entity.1.type"] == "workflow.langchain"
-        assert span_attributes["entity.2.name"] == "OpenSearchVectorSearch"
-        assert span_attributes["entity.2.type"] == "vectorstore.OpenSearchVectorSearch"
-        assert "entity.2.deployment" in span_attributes
-        assert span_attributes["entity.3.name"] == "amazon.titan-embed-text-v1"
-        assert span_attributes["entity.3.type"] == "model.embedding.amazon.titan-embed-text-v1"
-
-    if 'span.type' in span_attributes and span_attributes["span.type"] == "inference":
-        # Assertions for all inference attributes
-        assert span_attributes["entity.1.name"] == "bedrock_workflow"
-        assert span_attributes["entity.1.type"] == "workflow.generic"
-        assert span_attributes["entity.2.type"] == "inference.aws_sagemaker"
-        assert "entity.2.inference_endpoint" in span_attributes
-        assert span_attributes["entity.3.name"] == "ai21.j2-mid-v1"
-        assert span_attributes["entity.3.type"] == "model.llm.ai21.j2-mid-v1"
-
-
-        # Assertions for metadata
-        span_input, span_output, span_metadata = span.events
-        assert "completion_tokens" in span_metadata.attributes
-        assert "prompt_tokens" in span_metadata.attributes
-        assert "total_tokens" in span_metadata.attributes
 
 
 # {
