+++ tests/integration/test_langchain_bedrock_sample.py
@@ -1,99 +1,105 @@
+import boto3
 import bs4
+import pytest
+from common.custom_exporter import CustomConsoleSpanExporter
 from langchain import hub
+from langchain.text_splitter import RecursiveCharacterTextSplitter
+from langchain_aws import ChatBedrockConverse
 from langchain_chroma import Chroma
+from langchain_community.document_loaders import WebBaseLoader
+from langchain_community.embeddings import OpenAIEmbeddings
 from langchain_core.output_parsers import StrOutputParser
 from langchain_core.runnables import RunnablePassthrough
-from langchain_aws import ChatBedrockConverse
-from langchain.text_splitter import RecursiveCharacterTextSplitter
+from opentelemetry.sdk.trace.export import BatchSpanProcessor
+
 from monocle_apptrace.instrumentation.common.instrumentor import setup_monocle_telemetry
-from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
-from langchain_community.embeddings import OpenAIEmbeddings
-from langchain_community.document_loaders import WebBaseLoader
-from monocle.tests.common.custom_exporter import CustomConsoleSpanExporter
-import boto3
 
-
-bedrock_runtime_client = boto3.client(service_name='bedrock-runtime', region_name='us-east-1')
 custom_exporter = CustomConsoleSpanExporter()
-setup_monocle_telemetry(
-    workflow_name="bedrock_rag_workflow",
-    span_processors=[BatchSpanProcessor(custom_exporter)],
-    wrapper_methods=[],
-)
 
-llm = ChatBedrockConverse(
-    client=bedrock_runtime_client,
-    model_id="ai21.jamba-1-5-mini-v1:0",
-    temperature=0.1,
-)
+@pytest.fixture(scope="module")
+def setup():
+    setup_monocle_telemetry(
+        workflow_name="bedrock_rag_workflow",
+        span_processors=[BatchSpanProcessor(custom_exporter)],
+        wrapper_methods=[],
+    )
 
-loader = WebBaseLoader(
-    web_paths=("https://lilianweng.github.io/posts/2023-06-23-agent/",),
-    bs_kwargs=dict(
-        parse_only=bs4.SoupStrainer(
-            class_=("post-content", "post-title", "post-header")
-        )
-    ),
-)
-docs = loader.load()
+@pytest.mark.integration()
+def test_langchain_bedrock_sample(setup):
+    bedrock_runtime_client = boto3.client(service_name='bedrock-runtime', region_name='us-east-1')
+    llm = ChatBedrockConverse(
+        client=bedrock_runtime_client,
+        model_id="ai21.jamba-1-5-mini-v1:0",
+        temperature=0.1,
+    )
 
-text_splitter = RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=200)
-splits = text_splitter.split_documents(docs)
+    loader = WebBaseLoader(
+        web_paths=("https://lilianweng.github.io/posts/2023-06-23-agent/",),
+        bs_kwargs=dict(
+            parse_only=bs4.SoupStrainer(
+                class_=("post-content", "post-title", "post-header")
+            )
+        ),
+    )
+    docs = loader.load()
 
-vectorstore = Chroma.from_documents(
-    documents=splits,
-    embedding=OpenAIEmbeddings(),
-)
+    text_splitter = RecursiveCharacterTextSplitter(chunk_size=1000, chunk_overlap=200)
+    splits = text_splitter.split_documents(docs)
 
-retriever = vectorstore.as_retriever()
+    vectorstore = Chroma.from_documents(
+        documents=splits,
+        embedding=OpenAIEmbeddings(),
+    )
 
-prompt = hub.pull("rlm/rag-prompt")
+    retriever = vectorstore.as_retriever()
+
+    prompt = hub.pull("rlm/rag-prompt")
 
 
-def format_docs(docs):
-    return "\n\n".join(doc.page_content for doc in docs)
+    def format_docs(docs):
+        return "\n\n".join(doc.page_content for doc in docs)
 
 
-rag_chain = (
-    {"context": retriever | format_docs, "question": RunnablePassthrough()}
-    | prompt
-    | llm
-    | StrOutputParser()
-)
+    rag_chain = (
+        {"context": retriever | format_docs, "question": RunnablePassthrough()}
+        | prompt
+        | llm
+        | StrOutputParser()
+    )
 
-query = "What is Task Decomposition?"
-result = rag_chain.invoke(query)
+    query = "What is Task Decomposition?"
+    result = rag_chain.invoke(query)
 
-print(result)
+    print(result)
 
-spans = custom_exporter.get_captured_spans()
-for span in spans:
-    span_attributes = span.attributes
-    if "span.type" in span_attributes and span_attributes["span.type"] == "retrieval":
-        # Assertions for all retrieval attributes
-        assert span_attributes["entity.1.name"] == "Chroma"
-        assert span_attributes["entity.1.type"] == "vectorstore.Chroma"
-        assert "entity.1.deployment" in span_attributes
-        assert span_attributes["entity.2.name"] == "text-embedding-ada-002"
-        assert span_attributes["entity.2.type"] == "model.embedding.text-embedding-ada-002"
+    spans = custom_exporter.get_captured_spans()
+    for span in spans:
+        span_attributes = span.attributes
+        if "span.type" in span_attributes and span_attributes["span.type"] == "retrieval":
+            # Assertions for all retrieval attributes
+            assert span_attributes["entity.1.name"] == "Chroma"
+            assert span_attributes["entity.1.type"] == "vectorstore.Chroma"
+            assert "entity.1.deployment" in span_attributes
+            assert span_attributes["entity.2.name"] == "text-embedding-ada-002"
+            assert span_attributes["entity.2.type"] == "model.embedding.text-embedding-ada-002"
 
-    if "span.type" in span_attributes and span_attributes["span.type"] == "inference":
-        # Assertions for all inference attributes
-        assert span_attributes["entity.1.type"] == "inference.azure_oai"
-        assert "entity.1.provider_name" in span_attributes
-        assert "entity.1.inference_endpoint" in span_attributes
-        assert span_attributes["entity.2.name"] == "gpt-3.5-turbo-0125"
-        assert span_attributes["entity.2.type"] == "model.llm.gpt-3.5-turbo-0125"
+        if "span.type" in span_attributes and span_attributes["span.type"] == "inference":
+            # Assertions for all inference attributes
+            assert span_attributes["entity.1.type"] == "inference.azure_oai"
+            assert "entity.1.provider_name" in span_attributes
+            assert "entity.1.inference_endpoint" in span_attributes
+            assert span_attributes["entity.2.name"] == "gpt-3.5-turbo-0125"
+            assert span_attributes["entity.2.type"] == "model.llm.gpt-3.5-turbo-0125"
 
-        # Assertions for metadata
-        span_input, span_output, span_metadata = span.events
-        assert "completion_tokens" in span_metadata.attributes
-        assert "prompt_tokens" in span_metadata.attributes
-        assert "total_tokens" in span_metadata.attributes
+            # Assertions for metadata
+            span_input, span_output, span_metadata = span.events
+            assert "completion_tokens" in span_metadata.attributes
+            assert "prompt_tokens" in span_metadata.attributes
+            assert "total_tokens" in span_metadata.attributes
 
-    if not span.parent and span.name == "langchain.workflow":  # Root span
-        assert span_attributes["entity.1.name"] == "langchain_app_1"
-        assert span_attributes["entity.1.type"] == "workflow.langchain"
+        if not span.parent and span.name == "langchain.workflow":  # Root span
+            assert span_attributes["entity.1.name"] == "langchain_app_1"
+            assert span_attributes["entity.1.type"] == "workflow.langchain"
 
 # {
 #     "name": "langchain_core.vectorstores.base.VectorStoreRetriever",
