+++ tests/metamodel/metachain/methods.py
@@ -1,4 +1,4 @@
-from monocle.tests.metamodel.metachain.entities.inference import (
+from metamodel.metachain.entities.inference import (
     INFERENCE,
 )
