+++ tests/unit/app_hosting_types_test.py
@@ -1,26 +1,34 @@
+import logging
 import os
 import unittest
-from monocle_apptrace.instrumentation.common.wrapper_method import WrapperMethod
+
+from common.dummy_class import DummyClass
+from common.mock_exporter import MockExporter
 from opentelemetry.sdk.trace.export import SimpleSpanProcessor
+
+from monocle_apptrace.instrumentation.common.constants import (
+    service_name_map,
+    service_type_map,
+)
 from monocle_apptrace.instrumentation.common.instrumentor import setup_monocle_telemetry
 from monocle_apptrace.instrumentation.common.wrapper import task_wrapper
-from monocle_apptrace.instrumentation.common.constants import service_type_map, service_name_map
-from monocle.tests.common.dummy_class import DummyClass
-from monocle.tests.common.test_exporter import TestExporter
+from monocle_apptrace.instrumentation.common.wrapper_method import WrapperMethod
+
+logger = logging.getLogger(__name__)
 
 class TestHandler(unittest.TestCase):
     test_span_exporter = None
+    instrumentor = None
 
     def setUp(self, methodName='runTest'):
         app_name = "test"
-        self.test_span_exporter = TestExporter()
-
-        setup_monocle_telemetry(
+        self.test_span_exporter = MockExporter()
+        self.instrumentor = setup_monocle_telemetry(
             workflow_name=app_name,
             span_processors=[SimpleSpanProcessor(self.test_span_exporter)],
             wrapper_methods=[
                 WrapperMethod(
-                    package="dummy_class",
+                    package="common.dummy_class",
                     object_name="DummyClass",
                     method="dummy_chat",
                     span_name="langchain.workflow",
@@ -29,6 +37,14 @@ class TestHandler(unittest.TestCase):
                 )
             ]
         )
+
+    def tearDown(self) -> None:
+        try:
+            if self.instrumentor is not None:
+                self.instrumentor.uninstrument()
+        except Exception as e:
+            print("Uninstrument failed:", e)
+        return super().tearDown()
 
     def test_codespaces(self):
         dummy_class_1 = DummyClass()
