+++ tests/unit/langchain_async_test.py
@@ -2,43 +2,36 @@ import json
 import logging
 import os
 import time
-
 import unittest
-from unittest.mock import ANY, MagicMock, patch
 from unittest import IsolatedAsyncioTestCase
+from unittest.mock import ANY, MagicMock, patch
 
 import requests
-from monocle.tests.common.embeddings_wrapper import HuggingFaceEmbeddings
-from monocle.tests.common.http_span_exporter import HttpSpanExporter
+from common.embeddings_wrapper import HuggingFaceEmbeddings
+from common.fake_list_llm import FakeListLLM
+from common.http_span_exporter import HttpSpanExporter
 from langchain.prompts import PromptTemplate
 from langchain.schema import StrOutputParser
 from langchain_community.vectorstores import faiss
 from langchain_core.runnables import RunnablePassthrough
+from opentelemetry import trace
+from opentelemetry.sdk.resources import SERVICE_NAME, Resource
+from opentelemetry.sdk.trace import TracerProvider
+from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
+
+from monocle_apptrace.instrumentation.common.constants import (
+    SESSION_PROPERTIES_KEY,
+)
 from monocle_apptrace.instrumentation.common.instrumentor import (
     MonocleInstrumentor,
     set_context_properties,
     setup_monocle_telemetry,
 )
-from monocle_apptrace.instrumentation.common.constants import (
-    SESSION_PROPERTIES_KEY,
-)
-from monocle_apptrace.instrumentation.common.wrapper_method import WrapperMethod
-from opentelemetry import trace
-from opentelemetry.sdk.resources import SERVICE_NAME, Resource
-from opentelemetry.sdk.trace import TracerProvider
-from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
 from monocle_apptrace.instrumentation.common.span_handler import SpanHandler
+from monocle_apptrace.instrumentation.common.wrapper_method import WrapperMethod
 
-from monocle.tests.common.fake_list_llm import FakeListLLM
-
-logger = logging.getLogger()
-logger.setLevel(logging.DEBUG)
-fileHandler = logging.FileHandler('../traces.txt', 'w')
-formatter = logging.Formatter('%(asctime)s [%(levelname)s] %(name)s: %(message)s')
-fileHandler.setFormatter(formatter)
-logger.addHandler(fileHandler)
 events = []
-
+logger = logging.getLogger(__name__)
 
 class Test(IsolatedAsyncioTestCase):
