+++ tests/unit/langchain_workflow_name_test.py
@@ -3,49 +3,45 @@ import json
 import logging
 import os
 import time
-
 import unittest
 from unittest.mock import ANY, MagicMock, patch
 
 import requests
-
-from monocle.tests.common.embeddings_wrapper import HuggingFaceEmbeddings
-from monocle.tests.common.http_span_exporter import HttpSpanExporter
+from common.dummy_class import DummyClass
+from common.embeddings_wrapper import HuggingFaceEmbeddings
+from common.fake_list_llm import FakeListLLM
+from common.http_span_exporter import HttpSpanExporter
 from langchain.prompts import PromptTemplate
 from langchain.schema import StrOutputParser
 from langchain_community.vectorstores import faiss
+from langchain_core.messages.ai import AIMessage
 from langchain_core.runnables import RunnablePassthrough
+from opentelemetry import trace
+from opentelemetry.sdk.resources import SERVICE_NAME, Resource
+from opentelemetry.sdk.trace import TracerProvider
+from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
+from parameterized import parameterized
+
 from monocle_apptrace.instrumentation.common.constants import (
+    AWS_LAMBDA_ENV_NAME,
+    AWS_LAMBDA_SERVICE_NAME,
     AZURE_APP_SERVICE_ENV_NAME,
     AZURE_APP_SERVICE_NAME,
     AZURE_FUNCTION_NAME,
     AZURE_FUNCTION_WORKER_ENV_NAME,
     AZURE_ML_ENDPOINT_ENV_NAME,
     AZURE_ML_SERVICE_NAME,
-    AWS_LAMBDA_ENV_NAME,
-    AWS_LAMBDA_SERVICE_NAME
 )
 from monocle_apptrace.instrumentation.common.instrumentor import (
     MonocleInstrumentor,
     set_context_properties,
     setup_monocle_telemetry,
 )
-
-from opentelemetry import trace
-from opentelemetry.sdk.resources import SERVICE_NAME, Resource
-from opentelemetry.sdk.trace import TracerProvider
-from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
-
-from monocle.tests.common.fake_list_llm import FakeListLLM
-from parameterized import parameterized
 from monocle_apptrace.instrumentation.common.span_handler import SpanHandler
+from monocle_apptrace.instrumentation.common.wrapper_method import WrapperMethod
+from monocle_apptrace.instrumentation.metamodel.langchain import _helper
 
-logger = logging.getLogger()
-logger.setLevel(logging.DEBUG)
-fileHandler = logging.FileHandler('../traces.txt', 'w')
-formatter = logging.Formatter('%(asctime)s [%(levelname)s] %(name)s: %(message)s')
-fileHandler.setFormatter(formatter)
-logger.addHandler(fileHandler)
+logger = logging.getLogger(__name__)
 
 class TestWorkflowEntityProperties(unittest.TestCase):
 
@@ -175,6 +171,52 @@ class TestWorkflowEntityProperties(unittest.TestCase):
             except Exception as e:
                 print("Uninstrument failed:", e)
 
+
+    def test_custom_methods(self):
+        app_name = "test"
+        wrap_method = MagicMock(return_value=3)
+        setup_monocle_telemetry(
+            workflow_name=app_name,
+            span_processors=[
+                    BatchSpanProcessor(HttpSpanExporter("https://localhost:3000/api/v1/traces"))
+                ],
+            wrapper_methods=[
+                WrapperMethod(
+                    package="dummy_class",
+                    object_name="DummyClass",
+                    method="dummy_method",
+                    span_name="langchain.workflow",
+                    wrapper_method=wrap_method()),
+
+            ])
+        dummy_class_1 = DummyClass()
+
+        dummy_class_1.dummy_method()
+        wrap_method.assert_called_once()
+
+    def test_llm_response(self):
+        trace.set_tracer_provider(TracerProvider())
+        tracer = trace.get_tracer(__name__)
+        span = tracer.start_span("foo", start_time=0)
+
+        message = AIMessage(
+            content = "",
+            response_metadata = {
+                'token_usage': {'completion_tokens': 58, 'prompt_tokens': 584, 'total_tokens': 642}
+            }
+        )
+        instance = MagicMock()
+        metadata_dict = _helper.update_span_from_llm_response(response=message, instance=instance)
+        span.add_event(name="metadata", attributes=metadata_dict)
+        event_found = False
+        for event in span.events:
+            if event.name == "metadata":
+                attributes = event.attributes
+                assert attributes["completion_tokens"] == 58
+                assert attributes["prompt_tokens"] == 584
+                assert attributes["total_tokens"] == 642
+                event_found = True
+        assert event_found, "META_DATA event with token usage was not found"
 
 if __name__ == '__main__':
     unittest.main()
